% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cobraPhaseI.R
\name{cobraPhaseI}
\alias{cobraPhaseI}
\title{Find a feasible solution.}
\usage{
cobraPhaseI(cobra)
}
\arguments{
\item{cobra}{an object of class COBRA, this is a (long) list containing all settings
     from \code{\link{cobraInit}}}
}
\value{
\code{cobra}, an object of class COBRA
}
\description{
Find a feasible solution using the COBRA optimizer phase I by searching new infill points
 with the help of RBF surrogate models
}
\author{
Wolfgang Konen, Samineh Bagheri, Patrick Koch, Cologne Univeristy of Applied Sciences
}
\seealso{
\code{\link{cobraPhaseII}}, \code{\link{cobraInit}}
}

