% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultDebugRBF.R
\name{defaultDebugRBF}
\alias{defaultDebugRBF}
\title{Default settings for debug visualization RBF (only for \code{d==2})}
\usage{
defaultDebugRBF()
}
\value{
\code{DEBUG_RBF} a list of the follwing elements:
   \item{active}{If set to TRUE then \code{\link{debugVisualizeRBF}} is called every \code{DEBUG_RBF$every} iterations}
   \item{overlayTrueZ}{If set to TRUE overlay the true objective function}
   \item{DO_SNAPSHOT}{do rgl.snapshot every \code{DEBUG_RBF$every} iteration and store it \cr in \code{sprintf("images.d/\%s-\%03d.png",cobra$fname,npts)}}
   \item{every}{Frequency of calling the \code{\link{debugVisualizeRBF}} function}
}
\description{
Sets default values for debug visualization RBF  of SACOBRA.
}
\seealso{
\code{\link{debugVisualizeRBF}}
}
