% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RbfInter.R
\name{distLine}
\alias{distLine}
\title{Euclidean distance of \code{x} to all \code{xp}}
\usage{
distLine(x, xp)
}
\arguments{
\item{x}{vector of dimension d}

\item{xp}{n points \eqn{x_i} of dimension d are arranged in (n x d) matrix \code{xp}.
If \code{xp} is a vector, it is interpreted as (n x 1) matrix, i.e. d=1.}
}
\value{
vector of length n, the Euclidean distances
}
\description{
Euclidean distance of \code{x} to a line of points \code{xp}
}
\details{
\code{distLine} is up to 40x faster than using \code{\link[stats]{dist}} and taking only the first row or 
column of the distance matrix returned.
}
