% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SACOBRA.R
\docType{data}
\name{DRCL}
\alias{DRCL}
\title{Distance Requirement Cycle, long version}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
DRCL
}
\description{
Distance Requirement Cycle, long version: c(0.3,0.05, 0.001, 0.0005,0.0)
}
\keyword{datasets}
