% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SACOBRA.R
\docType{data}
\name{DRCS}
\alias{DRCS}
\title{Distance Requirement Cycle, short version}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
DRCS
}
\description{
Distance Requirement Cycle, short version: c(0.001,0.0)
}
\keyword{datasets}
