% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnArchive.R
\name{fnArchiveFactory}
\alias{fnArchiveFactory}
\title{construct wrapper \code{fnArchiveF} for a function \code{fn}}
\usage{
fnArchiveFactory(fn)
}
\arguments{
\item{fn}{function with vector argument x, should return a vector (obj,g1,g2,...,gn)}
}
\value{
\code{fnArchiveF}, a function which is a wrapper for \code{fn} and stores as a side effect the argument x and 
       the return value of \code{fn} in soluArchive and funcArchive, resp. These archives can be 
       retrieved at any time with \cr
           \code{(environment(fnArchiveF))$getSoluArchive()}  and \cr
           \code{(environment(fnArchiveF))$getFuncArchive()}
}
\description{
construct wrapper \code{fnArchiveF} for a function \code{fn}
}
\keyword{and}
\keyword{funcArchive}
\keyword{internal}
\keyword{reset}
\keyword{soluArchive}
