% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SACOBRA.R
\name{getFbest}
\alias{getFbest}
\title{Return best objective function value}
\usage{
getFbest(cobra)
}
\arguments{
\item{cobra}{an object of class COBRA (see  \code{\link{cobraInit}})}
}
\value{
the original objective function value at the best feasible solution
}
\description{
Return the original objective function value at the best feasible solution
}
\details{
Note: We cannot take the best function value via \code{cobra$fn}, because this 
may be modified by \code{plog()} or others )
}
\seealso{
\code{\link{getXbest}}
}
