% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SACOBRA.R
\name{getXbest}
\alias{getXbest}
\title{Return best feasible solution in original space}
\usage{
getXbest(cobra)
}
\arguments{
\item{cobra}{an object of class COBRA (see  \code{\link{cobraInit}})}
}
\value{
the best feasible solution in original space
}
\description{
Return best feasible solution in original space
}
\seealso{
\code{\link{getFbest}}
}
