% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SACOBRA.R
\docType{data}
\name{intern.archive.env}
\alias{intern.archive.env}
\title{Archiving Environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
intern.archive.env
}
\description{
intern.archive.env is an independent environment where every evaluated point and its evaluation by the real function
are stored in ARCHIVE and ARCHIVEY. This archive stores different values to \code{cobra$A} and \code{cobra$Fres} 
often during dubugging and visualisation cases where the real function is evaluated very often for debugging purposes.
}
\keyword{datasets}
