% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RbfInter.R
\name{interpRBF}
\alias{interpRBF}
\title{Apply cubic or Gaussian RBF interpolation to new data for d>1.}
\usage{
interpRBF(x, rbf.model)
}
\arguments{
\item{x}{vector holding a point of dimension d}

\item{rbf.model}{trained RBF model (or set of models), see \code{\link{trainCubicRBF}}
                 or \code{\link{trainGaussRBF}}}
}
\value{
value \eqn{s(x)} of the trained model at x \cr
                 - or - \cr
                 vector \eqn{s_j(x)} with values for all trained models \eqn{j=1,...,m} at x
}
\description{
Apply cubic or Gaussian RBF interpolation to new data for d>1.
}
\author{
Wolfgang Konen (\email{wolfgang.konen@fh-koeln.de})
}
\seealso{
\code{\link{trainCubicRBF}}, \code{\link{predict.RBFinter}}
}

