% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RbfInter.R
\name{interpRBF_old}
\alias{interpRBF_old}
\title{Apply cubic or Gaussian RBF interpolation to new data for d>1.}
\usage{
interpRBF_old(x, rbf.model)
}
\arguments{
\item{x}{vector holding a point of dimension d}

\item{rbf.model}{trained RBF model (or set of models), see \code{\link{trainCubicRBF}}
                 or \code{\link{trainGaussRBF}}}
}
\value{
val      value s(x) of the trained model at x
                 - or -
                 vector \eqn{s_j(x)} with values for all trained models \eqn{j=1,...,m} at x
}
\description{
Same as \code{\link{interpRBF}}, but it uses \code{dist} instead of
\code{distLine} and it is 40x slower on data
d=124, n_model=69, npts=249, see \code{testRbfInter2.R}
}
\author{
Wolfgang Konen (\email{wolfgang.konen@fh-koeln.de})
}
\seealso{
\code{\link{trainCubicRBF}}, \code{\link{predict.RBFinter}}
}
\keyword{internal}

