% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SACOBRA.R
\name{inverseRescale}
\alias{inverseRescale}
\title{Inverse Rescaling}
\usage{
inverseRescale(x, cobra)
}
\arguments{
\item{x}{a vector in the rescaled input space (usually \eqn{[-1,1]^d})}

\item{cobra}{list from \code{\link{cobraInit}}, we need here
\describe{
   \item{originalL}{   a vector with lower bounds in original input space}
   \item{originalU}{   a vector with upper bounds in original input space}
   \item{newlower}{    a number, the rescaled lower bound for all dimensions}
   \item{newupper}{    a number, the rescaled upper bound for all dimensions}
}}
}
\value{
\code{z},      inverse rescaling of vector x
}
\description{
Scale vector x in rescaled space back to original space
}
\seealso{
\code{\link{forwardRescale}}
}
