% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyEquCons.R
\name{modifyMu}
\alias{modifyMu}
\title{Modify equality margin \eqn{\mu}}
\usage{
modifyMu(Cfeas, Cinfeas, Tfeas, currentEps, cobra)
}
\arguments{
\item{Cfeas}{counter feasible iterates}

\item{Cinfeas}{counter infeasible iterates}

\item{Tfeas}{threshold counts}

\item{currentEps}{current value for \eqn{\mu}}

\item{cobra}{list of class COBRA}
}
\value{
\code{currentEps}, the modified value for \eqn{\mu}
}
\description{
Modify equality margin \eqn{\mu}
}
\keyword{internal}
