% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SACOBRA.R
\name{plogReverse}
\alias{plogReverse}
\title{Inverse of \code{\link{plog}}}
\usage{
plogReverse(y, pShift = 0)
}
\arguments{
\item{y}{function argument}

\item{pShift}{shift}
}
\value{
\eqn{plog^{-1}(y)}
}
\description{
Inverse of \code{\link{plog}}
}
\seealso{
\code{\link{plog}}
}
