% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SACOBRA.R
\name{RandomStart}
\alias{RandomStart}
\title{Random start Algorithm}
\usage{
RandomStart(cobra)
}
\arguments{
\item{cobra}{an object of class COBRA (see  \code{\link{cobraInit}})}
}
\value{
\code{cobra}, an object of class COBRA from \code{\link{cobraInit}}, with modified elements
     \item{\code{xStart}}{ new starting point, either random or \code{cobra$xBest} }
     \item{\code{noProgressCount}}{  number of consecutive iterations without a progress. 
                                     Set to 0, if a random start point is choosen. }
}
\description{
Return a cobra object with cobra$xStart adjusted
}
\keyword{internal}
