% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SACOBRA.R
\name{rescaleWrapper}
\alias{rescaleWrapper}
\title{Return a rescaled function}
\usage{
rescaleWrapper(fn, lower, upper, dimension, newlower, newupper)
}
\arguments{
\item{fn}{function with argument \code{x} to be rescaled}

\item{lower}{a vector with lower bounds in original input space}

\item{upper}{a vector with lower bounds in original input space}

\item{dimension}{length of vector \code{lower} and \code{upper}}

\item{newlower}{a number, the rescaled lower bound for all dimensions}

\item{newupper}{a number, the rescaled upper bound for all dimensions}
}
\value{
\code{newfn},      rescaled version of function \code{fn}
}
\description{
Return a rescaled function
}
\seealso{
\code{\link{forwardRescale}}, \code{\link{inverseRescale}}
}
