% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainSurrogates.R
\name{trainSurrogates}
\alias{trainSurrogates}
\title{Training surrogates for objective and constraint functions}
\usage{
trainSurrogates(cobra)
}
\arguments{
\item{cobra}{an object of class COBRA, this is a (long) list containing all settings
from \code{\link{cobraPhaseII}}}
}
\value{
\code{cobra}, an object of class COBRA, 
   with the following elements modified:
     \item{\code{constraintSurrogates}}{  surrogate models for all constraints }
     \item{\code{fitnessSurrogate}}{  surrogate model for objective function  }
     \item{\code{fitnessSurrogate1}}{ surrogate model for objective function,  w/o \code{\link{plog}} }
     \item{\code{fitnessSurrogate2}}{ surrogate model for objective function, with \code{\link{plog}} }
}
\description{
This function is called during the second phase of SACOBRA \code{\link{cobraPhaseII}} to train the surrogate models for objective and constarint functions 
of the optimization problem passed to the SACOBRA framework. This function takes all the so-far evalauted points stored in \code{cobra$A} and builds
}
\keyword{internal}
