% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyEquCons.R
\name{updateCobraEqu}
\alias{updateCobraEqu}
\title{Calculate cobra$xbest,fbest,ibest given the currently evaluated points.}
\usage{
updateCobraEqu(cobra, xNew)
}
\arguments{
\item{cobra}{an object of class COBRA (a long list)}

\item{xNew}{the last evaluated point}
}
\value{
\code{cobra}, an object of class COBRA, with potentially modified cobra$xbest, 
         cobra$fbest, cobra$ibest, cobra$noProgressCount
}
\description{
First, we calculate the set S of points being feasible in the inequalities and feasible
with margin cobra$currentEps in the equalities. If S is non-empty, take the point from S 
with best objective as cobra$xbest. If S is empty (there are no feasible points): If the new
point has lower maxViol than the maxViol of cobra$xbest, take it as the new cobra$xbest.
}
\details{
Called from updateSaveCobra.
}
\keyword{internal}
