% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobraPhaseI.R
\name{cobraPhaseI}
\alias{cobraPhaseI}
\title{Find a feasible solution.}
\usage{
cobraPhaseI(cobra)
}
\arguments{
\item{cobra}{an object of class COBRA, this is a (long) list containing all settings
from \code{\link{cobraInit}}}
}
\value{
\code{cobra}, an object of class COBRA
}
\description{
Find a feasible solution using the COBRA optimizer phase I by searching new infill points.
Please note that this phase can be skipped by setting the \code{cobra$skipPhaseI} parameter to \code{TRUE} in the initialization phase \code{\link{cobraInit}()}
}
\seealso{
\code{\link{cobraPhaseII}}, \code{\link{cobraInit}}
}
\author{
Wolfgang Konen, Samineh Bagheri, Patrick Koch, Cologne University of Applied Sciences
}
