% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobraPhaseI.R
\name{cobraPhaseI}
\alias{cobraPhaseI}
\title{Find a feasible solution.}
\usage{
cobraPhaseI(cobra)
}
\arguments{
\item{cobra}{an object of class COBRA, this is a (long) list containing all settings
from \code{\link{cobraInit}}}
}
\value{
\code{cobra}, an object of class COBRA from \code{\link{cobraInit}}, 
   enhanced here by the following elements (among others):
     \item{\code{df}}{  data frame with summary of the optimization run (see below)}
     \item{\code{A}}{ \code{(feval x d)}-matrix containing all evaluated points 
           in input space. If rescale==TRUE, all points are in \strong{rescaled} input space. }
     \item{\code{Fres}}{ a vector of the objective values of all evaluated points }
     \item{\code{Gres}}{ a \code{(feval x m)}-matrix of the constraint values of all evaluated points }
     \item{\code{predC}}{ a \code{(feval x m)}-matrix with the prediction of  
           \code{cobra$constraintSurrogates} at all evaluated points  }
     \item{\code{fbest}}{ the best feasible objective value found }
     \item{\code{xbest}}{ the point in input space yielding the best feasible objective value }
     \item{\code{ibest}}{ the corresponding iteration number (row of cobra$df, of cobra$A)}
 
  The data frame \strong{cobra$df} contains one row per iteration with columns 
  \describe{
     \item{iter}{ iteration index }
     \item{y}{   true objective value \code{Fres} }
     \item{predY}{  surrogate objective value. NA for the initial design points.}
     \item{predSolu}{  surrogate objective value at best-known solution \code{cobra$solu}, if given. 
           If \code{cobra$solu} is NULL, take the current point instead. NA for the initial design points.}
     \item{feasible}{ boolean indicating the feasibiltiy of infill point }
     \item{feasPred}{ boolean indicating if each infill point is feasible for \code{cobra$constraintSurrogates} }
     \item{nViolations}{ number of violated constraints }
     \item{maxViolation}{ maximum constraint violation. }
     \item{FEval}{  number of function evaluations in sequential optimizer. NA if it was a repair step }
     \item{Best}{  ever-best feasible objective value \code{fbest}. As long as there is 
           no feasible point, take among those with minimum number of violated constraints the
           one with minimum Fres. }
     \item{optimizer}{ e.g. "COBYLA"  }
     \item{optimizationTime}{  in sec}
     \item{conv}{ optimizer convergence code }
     \item{seed}{ the seed used in each run }
  }
}
\description{
Find a feasible solution using the COBRA optimizer phase I by searching new infill points.
This is done by minimizing the substitute function 
\deqn{ \mbox{Minimize}\quad  \sum_i{[\max(g_i(\vec{x}),0)]^2} , \vec{x} \in [\vec{a},\vec{b}] \subset \mathbf{R}^d }
until a feasible solution is found. This is for the case without equality constraints. 
If equality constraints are present, the above sum is extended by 
\deqn{   \sum_j{[h_j(\vec{x})-\mu]^2 + [-h_j(\vec{x})-\mu]^2}  \mbox{ with threshold } \mu}
}
\details{
\strong{Note:} Phase I can be skipped by setting the \code{cobra$skipPhaseI} parameter to \code{TRUE} in the initialization phase \code{\link{cobraInit}()}.
 
For further details on optimization with equality constraints see:\cr
\itemize{
  \item Bagheri, S.; Konen, W.; Baeck, T.:"Equality constraint handling for surrogate-assisted constrained optimization".
In: Tan, K.C. (ed.), WCCI'2016, Vancouver, 1924-1931, IEEE, 2016,
\url{http://www.gm.fh-koeln.de/~konen/Publikationen/Bagh16-WCCI.pdf}. 
   \item Bagheri, S.; Konen, W.; Baeck, T.: "How to Solve the Dilemma of Margin-Based Equality Handling Methods". 
In: Hoffmann, F. & Huellermeier, E. (Eds.),  Proceedings 28. Workshop Computational Intelligence, 
Universitaetsverlag Karlsruhe, 2018, 257-270, \strong{Young Author Award}, 
\url{https://blogs.gm.fh-koeln.de/ciop/files/2018/12/GMA2018.pdf} 
}
}
\seealso{
\code{\link{cobraPhaseII}}, \code{\link{cobraInit}}
}
\author{
Wolfgang Konen, Samineh Bagheri, Patrick Koch, Cologne University of Applied Sciences
}
