% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RbfInter.R
\name{interpRBF}
\alias{interpRBF}
\title{Apply the trained cubic, MQ or Gaussian RBF interpolation to new data for d>1.}
\usage{
interpRBF(x, rbf.model)
}
\arguments{
\item{x}{vector holding a point of dimension d}

\item{rbf.model}{trained RBF model (or set of models), see \code{\link{trainCubicRBF}} 
or \code{\link{trainGaussRBF}}}
}
\value{
value \eqn{s(\vec{x})} of the trained model at \eqn{\vec{x}} \cr
                 - or - \cr
                 vector \eqn{s_j( \vec{x})} with values for all trained models \eqn{j=1,...,m} at \eqn{\vec{x}}
}
\description{
Apply the trained cubic, MQ or Gaussian RBF interpolation to new data for d>1.
}
\seealso{
\code{\link{trainCubicRBF}},  \code{\link{trainMQRBF}},  \code{\link{trainGaussRBF}}, \code{\link{predict.RBFinter}}
}
\author{
Wolfgang Konen (\email{wolfgang.konen@th-koeln.de})
}
