% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultSAC.R
\name{setOpts}
\alias{setOpts}
\title{Merge the options from a partial list and the default list}
\usage{
setOpts(opts, defaultOpt)
}
\arguments{
\item{opts}{a partial list of options}

\item{defaultOpt}{a list with default values for every element}
}
\value{
a list combined from \code{opts} and \code{defaultOpt} where every available element 
   in \code{opts} overrides the default. For the rest of the elements the value from \code{defaultOpt}
   is taken. \cr
   A warning is issued for every element appearing in \code{opts} but not in \code{defaultOpt}
}
\description{
Merge the options from a partial list and the default list
}
\seealso{
\code{\link{defaultRI}}, \code{\link{defaultSAC}}, \code{\link{defaultTR}}, \code{\link{defaultEquMu}}
}
\author{
Samineh Bagheri, Wolfgang Konen, Cologne University of Applied Sciences
}
