% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSaveCobra.R
\name{updateSaveCobra}
\alias{updateSaveCobra}
\title{Update and save cobra}
\usage{
updateSaveCobra(
  cobra,
  ev1,
  subMin,
  sigmaD,
  penaF,
  gama,
  EPS,
  fitFuncPenalRBF,
  distRequirement,
  fitnessSurrogate = cobra$fitnessSurrogate
)
}
\arguments{
\item{cobra}{an object of class COBRA, this is a (long) list containing all settings
from \code{\link{cobraPhaseII}}}

\item{ev1}{a list filled by calls to \code{\link{evalReal}}. We need here the elements xNew,
feas, feasPred, feval, optimizerConvergence, optimizationTime, predY, predVal}

\item{subMin}{see \code{\link{cobraPhaseII}}}

\item{sigmaD}{see \code{\link{cobraPhaseII}}}

\item{penaF}{see \code{\link{cobraPhaseII}}}

\item{gama}{see \code{\link{cobraPhaseII}}}

\item{EPS}{see \code{\link{cobraPhaseII}}}

\item{fitFuncPenalRBF}{helper function from \code{\link{cobraPhaseII}}}

\item{distRequirement}{helper function from \code{\link{cobraPhaseII}}}

\item{fitnessSurrogate}{the model used for \code{predSoluFunc}}
}
\value{
\code{cobra}, an object of class COBRA, 
   enhanced here by the following elements (among others):
     \item{\code{df}}{  data frame with summary of the optimization run (see below)}
     \item{\code{df2}}{  data frame with additional summary information (see below)}
     \item{\code{dftr}}{  data frame with additional summary information for TR(see \code{\link{cobraPhaseII}})}
     \item{\code{fbest}}{ the best feasible objective value found }
     \item{\code{xbest}}{ the point in input space yielding the best feasible objective value }
     \item{\code{ibest}}{ the corresponding iteration number (row of \code{cobra$df}, of \code{cobra$A}) }
     \item{\code{fbestArray}}{ vector of all \code{fbest} }
     \item{\code{xbestArray}}{ vector of all \code{xbest} }
     
  The data frame \strong{cobra$df} contains one row per iteration with columns 
  \describe{
     \item{iter}{ iteration index }
     \item{y}{   true objective value \code{Fres} }
     \item{predY}{  surrogate objective value. Note: The surrogate may be trained on  
           plog-transformed training data, but \code{predY} is transformed back to the original 
           objective range. NA for the initial design points.}
     \item{predSolu}{  surrogate objective value at best-known solution \code{cobra$solu}, if given. 
           If \code{cobra$solu} is NULL, take the current point instead. Note: The surrogate may be trained on  
           plog-transformed training data, but \code{predSolu} is transformed back to the original 
           objective range. NA for the initial design points.}
     \item{feasible}{ boolean indicating the feasibiltiy of infill point }
     \item{feasPred}{ boolean indicating if each infill point is feasible for \code{cobra$constraintSurrogates} }
     \item{nViolations}{ number of violated constraints }
     \item{maxViolation}{ maximum constraint violation. }
     \item{FEval}{  number of function evaluations in sequential optimizer. NA if it was a repair step }
     \item{Best}{  ever-best feasible objective value \code{fbest}. As long as there is 
           no feasible point, take among those with minimum number of violated constraints the
           one with minimum Fres. }
     \item{optimizer}{ e.g. "COBYLA"  }
     \item{optimizationTime}{  in sec}
     \item{conv}{ optimizer convergence code }
     \item{dist}{ distance of the current point (row of \code{cobra$A}) to the true solution 
           \code{cobra$solu} in rescaled space. If there is more than one solution, take the one
           which has the minimum distance element (since this is the solution to which the 
           current run converges). }
     \item{distOrig}{ same as \code{dist}, but in original space  }
     \item{XI}{  the DRC element used in the current iteration }
     \item{seed}{ the seed used in each run }
  }
  The columns \code{feasPred, nViolations, maxViolation} are only present for constrained
  problems (i.e. if \code{cobra$CONSTRAINED=T}).

  The data frame \strong{cobra$df2} contains one row per phase-II-iteration with columns 
  \describe{
     \item{iter}{ iteration index}
     \item{predY}{  surrogate objective value. Note: The surrogate may be trained on  
           plog-transformed training data, but \code{predY} is transformed back to the original 
           objective range. NA for the initial design points.}
     \item{predVal}{   surrogate objective value + penalty }
     \item{predSolu}{   surrogate objective value at true solution (see \code{cobra$df$predSolu}) }
     \item{predSoluPenal}{   surrogate objective value + penalty at true solution (only diagnostics)}
     \item{sigmaD}{ the sigmaD element used in the current iteration (see \code{\link{cobraInit}})  }
     \item{penaF}{ penalty factor used in the current iteration (see \code{\link{cobraInit}}) }
     \item{XI}{  the DRC element used in the current iteration }
     \item{EPS}{ the current used margin for constraint function modeling (see \code{epsilonInit} in \code{\link{cobraInit}} ) }
  }
}
\description{
Helper for \code{\link{cobraPhaseII}}: make some assertion tests, 
update elements in object \code{cobra}, including data frames \code{df} and \code{df2}, 
and - if \code{cobra$saveIntermediate==TRUE} - save cobra in subdir \code{results/}.
Most importantly \code{cobra$xbest, cobra$fbest, cobra$ibest} are updated. They characterize the 
best feasible point (least violating point if no feasible point was found so far) and
influence the next starting point.
}
\details{
Note: the elements \code{A, Fres, Gres} of \code{cobra} are set in \code{updateInfoAndCounters},
an internal function of \code{\link{cobraPhaseII}}.
}
\seealso{
\code{\link{cobraPhaseII}}, \code{\link{cobraInit}}
}
